+++ src/js/video/video_clippy.js
@@ -1,4 +1,5 @@
-import { escapeHtml, NOMO_DEBUG, NOMO_WARN } from "js/lib.js";
+import { escapeHtml, NOMO_DEBUG, NOMO_WARN } from "js/lib/lib.js";
+import { sanitizeUrl } from "js/lib/sanitizeurl.ts";
 import { VideoBase } from "js/video/video_common.js";
 
 const ClippyLogo = `<svg width="18px" height="18px" viewBox="0 0 25 40" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M12.5 40C5.54688 40 0 34.4531 0 27.5V9.21875C0 4.14062 4.0625 0 9.21875 0C14.2969 0 18.4375 4.14062 18.4375 9.21875V25.625C18.4375 28.9062 15.7031 31.5625 12.5 31.5625C9.21875 31.5625 6.5625 28.9062 6.5625 25.625V12.5C6.5625 11.3281 7.5 10.3125 8.75 10.3125C9.92188 10.3125 10.9375 11.3281 10.9375 12.5V25.625C10.9375 26.5625 11.5625 27.1875 12.5 27.1875C13.3594 27.1875 14.0625 26.5625 14.0625 25.625V9.21875C14.0625 6.5625 11.875 4.375 9.21875 4.375C6.48438 4.375 4.375 6.5625 4.375 9.21875V27.5C4.375 32.0312 7.96875 35.625 12.5 35.625C16.9531 35.625 20.625 32.0312 20.625 27.5V12.5C20.625 11.3281 21.5625 10.3125 22.8125 10.3125C23.9844 10.3125 25 11.3281 25 12.5V27.5C25 34.4531 19.375 40 12.5 40Z" fill="black"></path></svg>`;
@@ -32,7 +33,10 @@ export class VideoClippy extends VideoBase {
                 NOMO_DEBUG("clippy json", json);
                 if(json.data){
                     if(json.data.cfVideoThumbnail){
-                        this.updateThumbnail(json.data.cfVideoThumbnail);
+                        let thumbnailUrl = sanitizeUrl(json.data.cfVideoThumbnail);
+                        if(thumbnailUrl !== "about:blank"){
+                            this.updateThumbnail(thumbnailUrl);
+                        }
                     }
                     if(json.data.title){
                         let newTitle = escapeHtml(json.data.title) + " - CLIPPY";